Add files via upload
A lovely new tutorial that reads in the big 3D-HST master catalog, reads in Arjen vanderWel's complementary catalog of sersic indices and effective radii, filters on galaxy redshift and stellar mass, and plots the results against my measurements for lensed galaxy s1110.  This is a nice tutorial on how to read and filter big catalogs with Pandas.